Hardware
Uploading the work I have done with the Arduino modules so far:

Using the RTC module (DS3231) I have been able to output the date and time onto a 16x2 LCD screen. 

I have then made a test function which creates an alarm for n seconds from current time - in the terminal it will state this and count down n seconds from current time. 

Using this, when the time meets the alarm will be fired. This will output noise from the buzzer which currently is connected to pin 8 currently.